Reorganise file & namespace structure of the QueryBuilder files
