+++ Polyel/src/Database/Query/Support/SqlCompile.trait.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Polyel\Database\Support;
+namespace Polyel\Database\Query\Support;
 
 trait SqlCompile
 {
